+++ src/types/index.ts
@@ -13,7 +13,7 @@ export type Panel =
   | 'variables'
   | 'notes'
   | 'settings'
-  | 'map'
+  // | 'map' // automapper disabled
   | 'alloc'
   | 'currency'
   | 'who'
@@ -25,7 +25,7 @@ export type PinnablePanel =
   | 'chat'
   | 'counter'
   | 'notes'
-  | 'map'
+  // | 'map' // automapper disabled
   | 'alloc'
   | 'currency'
   | 'who'